+++ src/main/java/org/apidesign/WebSocket.java
@@ -1,11 +1,12 @@
-package org.apidesign.polyfill;
+package org.apidesign;
 
 import java.io.File;
 import java.io.IOException;
-import java.util.Arrays;
 import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
+import org.apidesign.polyfill.Polyfill;
 import org.apidesign.polyfill.crypto.CryptoPolyfill;
 import org.apidesign.polyfill.timers.TimersPolyfill;
 import org.apidesign.polyfill.websocket.WebSocketPolyfill;
@@ -13,14 +14,26 @@ import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Source;
 import org.graalvm.polyglot.io.IOAccess;
 
-public class Main {
+public class WebSocket {
 
-    private Main() {
+    private WebSocket() {
+    }
+
+    public static void initializePolyfill(Context ctx, ExecutorService executor) {
+        Polyfill[] components = new Polyfill[]{
+            new TimersPolyfill(executor),
+            new CryptoPolyfill(),
+            new WebSocketPolyfill()
+        };
+
+        for (Polyfill component : components) {
+            component.initialize(ctx);
+        }
     }
 
     public static void main(String[] args) throws Exception {
         var path = "/all-y-websocket.js";
-        var demo = Main.class.getResource(path);
+        var demo = WebSocket.class.getResource(path);
         if (demo == null) {
             throw new IOException("Cannot find " + path);
         }
@@ -38,17 +51,11 @@ public class Main {
             var demoJs = Source.newBuilder("js", demo)
                     .mimeType("application/javascript+module")
                     .build();
-            var components = new Polyfill[]{
-                new TimersPolyfill(executor),
-                new CryptoPolyfill(),
-                new WebSocketPolyfill()
-            };
 
             CompletableFuture
                     .supplyAsync(b::build, executor)
                     .thenAcceptAsync(ctx -> {
-                        Arrays.stream(components).forEach(c -> c.initialize(ctx));
-
+                        initializePolyfill(ctx, executor);
                         ctx.eval(demoJs);
                     }, executor)
                     .get();
