+++ src/test/java/org/apidesign/WebSocketPolyfillTest.java
@@ -1,11 +1,9 @@
-package org.apidesign.polyfill.websocket;
+package org.apidesign;
 
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 
-import org.apidesign.polyfill.crypto.CryptoPolyfill;
-import org.apidesign.polyfill.timers.TimersPolyfill;
 import org.graalvm.polyglot.Context;
 import org.graalvm.polyglot.Source;
 import org.graalvm.polyglot.io.IOAccess;
@@ -32,15 +30,12 @@ public class WebSocketPolyfillTest {
         }
         executor = Executors.newSingleThreadExecutor();
         ctx = CompletableFuture
-                    .supplyAsync(() -> b.build(), executor)
-                    .thenApplyAsync(ctx -> {
-                        new TimersPolyfill(executor).initialize(ctx);
-                        new CryptoPolyfill().initialize(ctx);
-                        new WebSocketPolyfill().initialize(ctx);
-
-                        return ctx;
-                    }, executor)
-                    .get();
+                .supplyAsync(() -> {
+                    var context = b.build();
+                    WebSocket.initializePolyfill(context, executor);
+                    return context;
+                }, executor)
+                .get();
     }
 
     @AfterClass
